% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{addchr}
\alias{addchr}
\alias{addchr,GRanges-method}
\alias{addchr,loops-method}
\title{Add 'chr' to GRanges seqnames}
\usage{
addchr(dlo)

\S4method{addchr}{loops}(dlo)

\S4method{addchr}{GRanges}(dlo)
}
\arguments{
\item{dlo}{A loops object or GRanges object}
}
\value{
An identical loops object or GRanges object 'chr' added
}
\description{
\code{addchr} takes a loops object or GRanges object and 
simply adds 'chr' to seqnames
}
\details{
Often times, performing functions on GRanges objects can go awry 
if the seqnames are systematically different. A common example of
this is when some GRanges objects has the format of 'chr1' 
while the other has '1'. We can add 'chr' to the first object
}
\examples{
library(GenomicRanges)
regA <- GRanges(c('1'),ranges=IRanges(c(36200000),c(36300000)))
addchr(regA)
regA
rmchr(regA)
regA
}

