% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{methods}
\name{annotateAnchors.bigwig}
\alias{annotateAnchors.bigwig}
\alias{annotateAnchors.bigwig,ANY-method}
\title{Add meta data column to anchors based on .bigwig}
\usage{
annotateAnchors.bigwig(dlo, file, FUN = mean, pad = 0)

\S4method{annotateAnchors.bigwig}{ANY}(dlo, file, FUN = mean, pad = 0)
}
\arguments{
\item{dlo}{A loops object whose anchors will be annotated}

\item{file}{A Granges object corresponding to locations of interest}

\item{FUN}{A function used to combine multiple values observed in a single anchor; default is mean}

\item{pad}{An integer value of to pad the anchors of the loops object; default is 0}
}
\value{
A loops object with new numeric meta data column in anchors
}
\description{
\code{annotateAnchors.bigwig} adds a numeric variable to meta data 
columns in the anchors slot based on a user-specified .bigwig
file.
}
\details{
This function adds a meta data column to anchors of the specified
loops object. All values from the .bigwig file that overlap with the 
each anchor are handled by the FUN (default is to average them) to 
produce a single value added to the mcols of the anchors.
}
\examples{
# Annotate whether anchors are near a gene body; within 1kb
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
gb <-getHumanGenes()
loops.small <- annotateAnchors(loops.small,gb,'nearGeneBody')

}

