% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{methods}
\name{annotateAnchors}
\alias{annotateAnchors}
\alias{annotateAnchors,loops,GRanges,character,missing-method}
\alias{annotateAnchors,loops,GRanges,character,numeric-method}
\title{Add meta data column to anchors based on GRanges}
\usage{
annotateAnchors(dlo, features, featureName, maxgap)

\S4method{annotateAnchors}{loops,GRanges,character,missing}(dlo, features,
  featureName, maxgap = 1000)

\S4method{annotateAnchors}{loops,GRanges,character,numeric}(dlo, features,
  featureName, maxgap)
}
\arguments{
\item{dlo}{A loops object whose anchors will be annotated}

\item{features}{A Granges object corresponding to locations of interest}

\item{featureName}{A string that will be the mcol name in anchors}

\item{maxgap}{A value of max permissible gap between a feature and anchor}
}
\value{
A loops object with new meta data column in anchors
}
\description{
\code{annotateAnchors} adds a logical variable to meta data columns in the
anchors based on a GRanges object of features' genomic coordinates
}
\details{
This function adds column of TRUE/FALSE values on the loops object
anchors whether a feature is observed nearby in \code{features}. The name
of this column that will be in the anchors GRanges object is specified by
a user defined string \code{featureName}. Gap tolerance between a feature
and an anchor is specified by \code{maxgap}, where the default is 1,000bp.
}
\examples{
# Annotate whether anchors are near a gene body; within 1kb
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
gb <-getHumanGenes()
loops.small <- annotateAnchors(loops.small,gb,'nearGeneBody')

# Adding close to gene bodies with no gap tolerance
loops.small <- annotateAnchors(loops.small,gb,'inGeneBody',0)

}

