% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{methods}
\name{annotateLoops.dge}
\alias{annotateLoops.dge}
\alias{annotateLoops.dge,ANY-method}
\title{Annotate enhancer-promoter loops with differential gene expression}
\usage{
annotateLoops.dge(lto, deseq_res, multiple = FALSE)

\S4method{annotateLoops.dge}{ANY}(lto, deseq_res, multiple = FALSE)
}
\arguments{
\item{lto}{A loops object whose loops will be annotated}

\item{deseq_res}{A data.frame}

\item{multiple}{Annotate loops with multiple TSS? Default = FALSE}
}
\value{
A loops object if multiple = FALSE or data frame if multiple = TRUE
}
\description{
\code{annotateLoops.dge} adds columns to the rowData slot of a loops
object that shows summary statistics corresponding TSS of a gene name 
based on the genes.tss rowData column. This function should be used
following the \code{keepEPloops} function.
}
\details{
This function links enhancer-promoter loops and differential gene
expression results. The rownames of the deseq_res slots should correspond
to the gene names in the gene.tss column of the rowData slot of the loops
object. The function returns a loops object if multiple is specified as FALSE
which is the case by default. Otherwise, if multiple is TRUE, then this
function returns a data frame since each loop may have more than moe TSS. 
One can reproduce this dataframe quickly when multiple = FALSE using the
summary() function on the returned loops object.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
h3k27ac_j <- system.file('extdata','Jurkat_H3K27ac_chr1.narrowPeak',package='diffloop')
h3k27ac <- rmchr(padGRanges(bedToGRanges(h3k27ac_j), pad = 1000))
promoter <- padGRanges(getHumanTSS(c('1')), pad = 1000)
#small.ep <- keepEPloops(loops.small, h3k27ac, promoter)
#ADD SOMETHING HERE.

}

