% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{bedToGRanges}
\alias{bedToGRanges}
\alias{bedToGRanges,character-method}
\title{Read a file and make a GRanges object}
\usage{
bedToGRanges(file)

\S4method{bedToGRanges}{character}(file)
}
\arguments{
\item{file}{A string specifiying .bed file location}
}
\value{
A GRanges object
}
\description{
\code{bedToGRanges} takes a string corresponding to a file 
and creates a GRanges object, retaining meta-data
}
\details{
Useful function to read in a .bed file to create a GRanges object
where the meta-data is presevered. Useful for later functions like
\code{annotateAnchors}
}
\examples{
#Read in CTCF Jurkat peaks in
ctcf_j <- system.file('extdata','Jurkat_CTCF_chr1.narrowPeak',package = 'diffloop')
ctcf <- bedToGRanges(ctcf_j)


}

