% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\docType{methods}
\name{calcLDSizeFactors}
\alias{calcLDSizeFactors}
\alias{calcLDSizeFactors,loops-method}
\title{Compute normalizing factors for each sample}
\usage{
calcLDSizeFactors(dlo)

\S4method{calcLDSizeFactors}{loops}(dlo)
}
\arguments{
\item{dlo}{A loops object with unnormalized size factors}
}
\value{
A loops object with new size factors in \code{colData}
}
\description{
\code{calcLDSizeFactors} takes a loops object computes size 
factors based for each sample
}
\details{
This function updates the \code{loops} object with new
\code{sizeFactor} values for
each sample in the \code{colData} slot using a method identical to that
employed in \code{DESeq2.}
}
\examples{
# Computing normalizing factors from the full ChIA-PET Data
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
loops.small <- calcLDSizeFactors(loops.small)
}

