% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruan.R
\docType{methods}
\name{callCCDs}
\alias{callCCDs}
\alias{callCCDs,ANY-method}
\title{Compute Chromatin Contact Domains (CCDs)}
\usage{
callCCDs(lo, petWeights = FALSE, lowCoveragePercentile = 0.05)

\S4method{callCCDs}{ANY}(lo, petWeights = FALSE,
  lowCoveragePercentile = 0.05)
}
\arguments{
\item{lo}{A loops object}

\item{petWeights}{Boolean to weight loop coverage by number of PETs. Default = FALSE}

\item{lowCoveragePercentile}{Percentile of low coverage to be dropped. Default = 0.05}
}
\value{
A GRanges object of called Chromatin Contact domains
}
\description{
\code{callCCDs} determines regions
}
\details{
This funciton returns a GRanges object of regions determined to be Chromatin Contact
Domains as defined in the Tang et al. 2015 paper from the Ruan group. Users
can choose to weight the loops by the total number PETs (across all samples)
or not and what percent. For details of this method, see page 12 of the supplement
of PMID:26686651. Make sure there are only loops within a chromosome before calling this.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
lo <- subsetLoops(loops.small, c(1,2,5,6,7,8,9,27,69))
ccd <- callCCDs(lo, petWeights = TRUE, lowCoveragePercentile = 0.5)

}

