% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruan.R
\docType{methods}
\name{computeBoundaryScores}
\alias{computeBoundaryScores}
\alias{computeBoundaryScores,loops-method}
\title{Compute boundary scores for genomic loci in between anchors}
\usage{
computeBoundaryScores(dlo, samples = 0, windowSize = 1e+06)

\S4method{computeBoundaryScores}{loops}(dlo, samples = 0,
  windowSize = 1e+06)
}
\arguments{
\item{dlo}{A loops object}

\item{samples}{= 0 Vector indexing which samples should be used. 0 is all}

\item{windowSize}{= 1000000}
}
\value{
A GRanges object with genomic loci and boundary scores in the mcols
}
\description{
\code{computeBoundaryScores} determines the
 boundary scores corresponding to the Genomic region either
in bewteen pairs of anchors. To achieve this, the number of PETs
for a set of samples (default is all = 0) is summed over a window
(default 1MB) on the left (A) and the right (B) of gap. Thus sum 
of the number of PETs in these windows is devided by the number 
of PETs that span the two loci, plus 1 (C). A larger value
corresponds to a stronger boundary.
}
\details{
Warning: this function is slow; there is a progress bar outputted to give 
an anticipated runntime.
}
\examples{
# Return the width for loops 
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
BS <- computeBoundaryScores(loops.small, samples = 0, windowSize = 1000000)

}

