% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{dim,loopdata-method}
\alias{dim,loopdata-method}
\title{See dimensions of loopdata}
\usage{
\S4method{dim}{loopdata}(x)
}
\arguments{
\item{x}{A loopdata object}
}
\value{
A data.frame of dimensions of the loopdata object,
including number of anchors, loops, samples, and column data
attributes
}
\description{
See dimensions of loopdata
}

