% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{dim,loops-method}
\alias{dim,loops-method}
\title{See dimensions of loops object}
\usage{
\S4method{dim}{loops}(x)
}
\arguments{
\item{x}{A loops object}
}
\value{
A data.frame of dimensions of the loops object,
including number of anchors, interactions, samples, and column data
attributes
}
\description{
See dimensions of loops object
}

