% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{dim,looptest-method}
\alias{dim,looptest-method}
\title{See dimensions of looptest}
\usage{
\S4method{dim}{looptest}(x)
}
\arguments{
\item{x}{A looptest object}
}
\value{
A data.frame of dimensons of the looptest object,
including number of anchors, loops, samples, column data
attributes, and columns in the results subobject
}
\description{
See dimensions of looptest
}

