% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\docType{methods}
\name{featureTest}
\alias{featureTest}
\alias{featureTest,loops,GRanges-method}
\title{Combined association test for all loops in a defined region}
\usage{
featureTest(x, features)

\S4method{featureTest}{loops,GRanges}(x, features)
}
\arguments{
\item{x}{A loops object}

\item{features}{A GRanges object defining regions for a combined test}
}
\value{
A data.frame sorted by FDR
}
\description{
\code{featureTest} takes a \code{loops} and 
genomic coordinates of regions and computes combined significance 
metrics for each region using the Simes procedure
}
\details{
This function returns a data.frame sorted by FDR of each region. Assumes
the region name is specified in the GRanges object with \code{id} column. 
Each feature is a one row in the GRanges object. The combined significance
measure per feature is computed via the Simes method for intrachromosomal 
loops where at least one anchor from the loop overlaps with the region of 
interest.
}
\examples{
# Human genes chromosome 1 regional association
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
assoc <- loopAssoc(loops.small, coef = 2)
# Gene based association
sw_jn <- featureTest(assoc, getHumanGenes(c('1')))
}

