% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{filterLoops}
\alias{filterLoops}
\alias{filterLoops,ANY-method}
\title{Filter loops}
\usage{
filterLoops(dlo, width = 0, nreplicates = 0, nsamples = 1)

\S4method{filterLoops}{ANY}(dlo, width = 0, nreplicates = 0, nsamples = 1)
}
\arguments{
\item{dlo}{A loopdata object}

\item{width}{Minimum loop width}

\item{nreplicates}{Minimum number of counts per loop}

\item{nsamples}{Minimum number of samples per loop per counts}
}
\value{
A loopdata object
}
\description{
\code{filterLoops} filters out loops that aren't wide, aren't prevalent
within samples or prevalent between samples
}
\details{
Function that restricts loops in a loopdata object. \code{width} specifies 
the minimum width between anchors. Default is zero. \code{nreplicates}
restricts loops to at least this specified amount of counts is present
in at least one sample. Instead of \code{nreplicates} being present in only
one sample, \code{nsamples} specifies how many individual samples that a
loop must have \code{nreplicates} in to be included after filtering.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
# Restrict loops to > 5kb width
filtered.jpn1 <- filterLoops(jpn_chr1reg, 5000, 0, 0)
# Restrict loops to > 5kb width and have >= 3 replicates in >= 1 sample
filtered.jpn2 <- filterLoops(jpn_chr1reg, 5000, 3, 1)
# Restrict loops to > 10kb width and have >= 3 replicates in >= 2 samples
filtered.jpn3 <- filterLoops(jpn_chr1reg, 10000, 3, 2)
}

