% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\docType{methods}
\name{filterSpanningLoops}
\alias{filterSpanningLoops}
\alias{filterSpanningLoops,loops,GRanges-method}
\title{Retain loops spanning some genomic feature}
\usage{
filterSpanningLoops(dlo, gf)

\S4method{filterSpanningLoops}{loops,GRanges}(dlo, gf)
}
\arguments{
\item{dlo}{A loops object}

\item{gf}{A GRanges object of features}
}
\value{
An loops object
}
\description{
\code{filterSpanningLoops} returns a loops object where the ends of the 
anchors completely span one or more genomic feature (e.g. boundary)
}
\details{
Rather than a simple overlap, the function by default requires a
genomic locus to be completely contacints
}
\examples{
# Return the width for loops 
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
w <- loopWidth(loops.small)

}

