% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo}
\alias{geneinfo}
\title{Human/mouse exon locations}
\format{A GRanges object 
\describe{
  \item{chrom}{Chromosomes without "chr"}
  \item{start}{exon start location}
  \item{stop}{exon end location}
  \item{gene}{Gene Name}
  \item{score}{dummy column there for sushi}
  \item{strand}{+1 or -1 to indicate side of DNA}
  ...
}}
\source{
biomaRt July 2015 stable build
}
\usage{
geneinfo
}
\value{
A data.frame
}
\description{
A dataframe used for plotting annotation for human
and mouse. Each loaded .rda has the same variable
called "geneinfo" (so don't co-load these), but 
the files differ by an m orh
}
\keyword{datasets}

