% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{methods}
\name{getHumanTSS}
\alias{getHumanTSS}
\alias{getHumanTSS,character-method}
\alias{getHumanTSS,missing-method}
\title{Get Human Transcription Start Sites}
\usage{
getHumanTSS(chr, cache = TRUE)

\S4method{getHumanTSS}{missing}(chr, cache = TRUE)

\S4method{getHumanTSS}{character}(chr, cache = TRUE)
}
\arguments{
\item{chr}{Specifies what chromosomes are desired for the TSS}

\item{cache}{logic variable (default = TRUE) to use TSS from July.2015 freeze}
}
\value{
A GRanges object
}
\description{
\code{getHumanTSS} returns a \code{GRanges} object of all 
transcription start sites for humans
}
\details{
This function returns a \code{GRanges} object with the coordinates and
gene TSS. The start and end of the IRanges slot will be the same number,
so consider using the \code{padGRanges} function after calling this function.
}
\examples{
# Grab all transition start sites genome-wide
human.TSS <- getHumanTSS()
}

