% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{methods}
\name{getMouseGenes}
\alias{getMouseGenes}
\alias{getMouseGenes,character-method}
\alias{getMouseGenes,missing-method}
\title{Get protein coding gene regions}
\usage{
getMouseGenes(chr)

\S4method{getMouseGenes}{missing}(chr)

\S4method{getMouseGenes}{character}(chr)
}
\arguments{
\item{chr}{A vector of chromosomes}
}
\value{
A GRanges object
}
\description{
\code{getMouseGenes} returns a \code{GRanges} object of all protein
coding genes genome-wide or within specified chromosomes. Annotation
is from regions from mm9 and protein coding genes.
}
\details{
This function returns a \code{GRanges} object with the coordinates and
gene IDs of all protein coding genes either genome-wide 
(by default) orspecified within a particular chromosome.
}
\examples{
# Grab all protein coding gene locations genome-wide
pc.genes <- getMouseGenes()
# Grab all protein coding gene loctions on chromosome 1
chr1 <- getHumanGenes(c('1'))
}

