% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{methods}
\name{getMouseTSS}
\alias{getMouseTSS}
\alias{getMouseTSS,character-method}
\alias{getMouseTSS,missing-method}
\title{Get Mouse Transcription Start Sites}
\usage{
getMouseTSS(chr)

\S4method{getMouseTSS}{missing}(chr)

\S4method{getMouseTSS}{character}(chr)
}
\arguments{
\item{chr}{Specifies what chromosomes are desired for the TSS#}
}
\value{
A GRanges object
}
\description{
\code{getMouseTSS} returns a \code{GRanges} object of all 
transcription start sites for humans. Regions from mm9 for 
protein coding regions.
}
\details{
This function returns a \code{GRanges} object with the coordinates and
gene TSS. The start and end of the IRanges slot will be the same number,
so consider using the \code{padGRanges} function after calling this function.
}
\examples{
# Grab all transition start sites genome-wide
mouse.TSS <- getMouseTSS()
}

