% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{head,loopdata-method}
\alias{head,loopdata-method}
\title{Extract first part of loopdata}
\usage{
\S4method{head}{loopdata}(x, n = 6, ...)
}
\arguments{
\item{x}{A looptest object}

\item{n}{Number of lines to view}

\item{...}{Other non-essential params}
}
\value{
A loopdata object
}
\description{
Extract first part of loopdata
}

