% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{head,loops-method}
\alias{head,loops-method}
\title{Extract first part of loops object}
\usage{
\S4method{head}{loops}(x, n = 6, ...)
}
\arguments{
\item{x}{A loops object}

\item{n}{Number of lines to view}

\item{...}{Other non-essential params}
}
\value{
A loops object
}
\description{
Extract first part of loops object
}

