% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{head,looptest-method}
\alias{head,looptest-method}
\title{Extract first part of looptest}
\usage{
\S4method{head}{looptest}(x, n = 6, ...)
}
\arguments{
\item{x}{A looptest object}

\item{n}{Number of lines to view}

\item{...}{Other non-essential params}
}
\value{
A looptest object
}
\description{
Extract first part of looptest
}

