% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.genes}
\alias{human.genes}
\title{Human protein coding genes}
\format{A GRanges object 
\describe{
  \item{seqnames}{Chromosomes without "chr"}
  \item{ranges}{start/end loci}
  \item{strand}{not specified ('*' everywhere)}
  \item{id}{Gene Name}
  ...
}}
\source{
biomaRt July 2015 stable build
}
\usage{
human.genes
}
\value{
A GRanges object
}
\description{
A GRanges object with the human protein-coding genes
}
\keyword{datasets}

