% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.TSS}
\alias{human.TSS}
\title{Human 60k+ transcription start sites}
\format{A GRanges object 
\describe{
  \item{seqnames}{Chromosomes without "chr"}
  \item{ranges}{start/end loci are same}
  \item{strand}{not specified ('*' everywhere)}
  \item{id}{Gene Name}
  ...
}}
\source{
biomaRt July 2015 stable build
}
\usage{
human.TSS
}
\value{
A GRanges object
}
\description{
A GRanges object with all loci of transcription start sites
}
\keyword{datasets}

