% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{interchromosomal}
\alias{interchromosomal}
\alias{interchromosomal,loopdata-method}
\title{Loops between chromosomes}
\usage{
interchromosomal(dlo)

\S4method{interchromosomal}{loopdata}(dlo)
}
\arguments{
\item{dlo}{A loopdata object}
}
\value{
A loopdata object with all loops on different chromosomes
}
\description{
\code{interchromosomal} restricts loops to those where anchors are observed
on different chromosomes
}
\details{
This function subsets the \code{loopdata} object into only those loops that
have anchors on different chromosomes
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)

# Compute number of loops on same chromosome
dim(intrachromosomal(jpn_chr1reg))
samechromo <- intrachromosomal(jpn_chr1reg)

# Compute number of loops on same chromosome
# dim(interchromosomal(jpn_chr1reg))
# This will throw and error since the toy only has intrachromosomal loops

}

