% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{intrachromosomal}
\alias{intrachromosomal}
\alias{intrachromosomal,loops-method}
\title{Loops within chromosomes}
\usage{
intrachromosomal(dlo)

\S4method{intrachromosomal}{loops}(dlo)
}
\arguments{
\item{dlo}{A loops object}
}
\value{
A loops object where all loops are on the same chromosome.
}
\description{
\code{intrachromosomal} restricts interactions to those where anchors are observed
on the same chromosomes
}
\details{
This function subsets the \code{loops} object into only those interactions that
have both anchors on the same chromosome
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)

# Compute number of interactions on same chromosome
dim(intrachromosomal(loops.small))
samechromo <- intrachromosomal(loops.small)

}

