% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jpn_chr1reg}
\alias{jpn_chr1reg}
\title{chr1:36000000-36300000 loops}
\format{A small loopdata object 
\describe{
  \item{anchors}{GRanges object of anchor locations}
  \item{loops}{indexes of interactions }
  \item{samples}{Two replicates each of jurkat, naive, and primed cells}
  \item{colData}{Groups identifying cell type and unnormalized sizeFactors}
  ...
}}
\source{
subsetRegion(full,GRanges(c('1'),IRanges(c(36000000),c(36300000))))
}
\usage{
jpn_chr1reg
}
\value{
A loopdata object
}
\description{
A loopdata object containing unique 108 loops with 27 anchors
for 6 samples and corresponding colData
}
\keyword{datasets}

