% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{methods}
\name{keepCTCFloops}
\alias{keepCTCFloops}
\alias{keepCTCFloops,loops,GRanges-method}
\title{Keep CTCF loops}
\usage{
keepCTCFloops(lto, ctcf)

\S4method{keepCTCFloops}{loops,GRanges}(lto, ctcf)
}
\arguments{
\item{lto}{A loops object whose loops will be annotated}

\item{ctcf}{GRanges object corresponding to locations of CTCF peaks}
}
\value{
A loops object with all loops having both anchors in the GRanges region
}
\description{
\code{keepCTCFloops} returns loops that are nearby CTCF peaks
as determined by some external data in a GRanges object
}
\details{
This function works similar to the \code{annotateLoops} function but
returns only CTCF loops that are defined in this function.
However, loops in \code{annotateLoops} may have a different annotation
based on their priority scheme. For example, an e-p loop from 
\code{annotateLoops} may be returned as a CTCF loop by this function
if the loop had both annotations. These peaks don't necessarily 
need to be CTCF peaks, so using a GRanges object with enhancers
or promoters to determine e-e loops and p-p loops could also
be used in this function
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
ctcf_j <- system.file('extdata','Jurkat_CTCF_chr1.narrowPeak',package='diffloop')
ctcf <- rmchr(padGRanges(bedToGRanges(ctcf_j), pad = 1000))
small.ctcf <- keepCTCFloops(loops.small, ctcf)

}

