% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\docType{methods}
\name{loopAssoc}
\alias{loopAssoc}
\alias{loopAssoc,loops-method}
\title{Generalized differential Loop Calling}
\usage{
loopAssoc(y, method = "edgeR", design = NULL, coef = NULL,
  contrast = NULL)

\S4method{loopAssoc}{loops}(y, method = "edgeR", design = NULL,
  coef = NULL, contrast = NULL)
}
\arguments{
\item{y}{A loops object for association}

\item{method}{Specifies association; either "Voom" or "edgeR"}

\item{design}{A design matrix of the samples; required for "Voom"}

\item{coef}{A vector for the coefficient of GLM. See edgeR manual}

\item{contrast}{A vector for the contrast. See edgeR manual}
}
\value{
A loops object
}
\description{
\code{loopAssoc} takes a \code{loops} object and prepares it for the
returns another \code{loops} object with summary statistics
per-loop in the \code{rowData}
}
\details{
By the default, we generate is to generate a design matrix from
\code{loops@colData$groups}. Currently, 'edgeR' and 'Voom' are the
two supported association  
methods, but new association tests may be added in later developments.
}
\examples{
# Differential loop fit
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
# assoc <- loopAssoc(loops.small, coef = 2)
}

