% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_io.R
\docType{methods}
\name{loopdataMake}
\alias{loopdataMake}
\alias{loopdataMake,character,character,missing-method}
\alias{loopdataMake,character,character,numeric-method}
\alias{loopdataMake,character,missing,missing-method}
\alias{loopdataMake,character,missing,numeric-method}
\title{Read preprocessed ChiA-PET data}
\usage{
loopdataMake(beddir, samples, mergegap)

\S4method{loopdataMake}{character,character,missing}(beddir, samples, mergegap)

\S4method{loopdataMake}{character,missing,missing}(beddir, samples, mergegap)

\S4method{loopdataMake}{character,character,numeric}(beddir, samples, mergegap)

\S4method{loopdataMake}{character,missing,numeric}(beddir, samples, mergegap)
}
\arguments{
\item{beddir}{A string. The preprocessed data directory}

\item{samples}{A character vector. Optional list of samples to read in}

\item{mergegap}{An integer value of the radius to merge anchors; default 0}
}
\value{
A loopdata object
}
\description{
\code{loopdataMake} reads in a data directory created by the
\code{dnaloop} preprocessing pipeline and returns a loopdata object
}
\details{
This function reads in preprocessed ChIA-PET data produced by the
\code{dnaloop} preprocessing pipeline. The preprocessed directory 
contains one subdirectory per sample. The \code{samples} argument specifies
which samples are read. if \code{samples} is not specified all samples will
be read.
}
\examples{
# Reading in all samples, no mergegap:
bd<- system.file('extdata', 'esc_jurkat', package='diffloopdata')
loops <- loopdataMake(bd)

# Reading in a subset of samples, 1kb mergegap:
samples <- c('naive_esc_1', 'naive_esc_2')
naive <- loopdataMake(bd, samples, 1000)

}

