% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\docType{methods}
\name{loopdataSubset}
\alias{loopdataSubset}
\alias{loopdataSubset,loopdata,loopdata-method}
\title{Subset two difloop objects}
\usage{
loopdataSubset(dlo1, dlo2)

\S4method{loopdataSubset}{loopdata,loopdata}(dlo1, dlo2)
}
\arguments{
\item{dlo1}{A loopdata object}

\item{dlo2}{A loopdata object}
}
\value{
A loopdata obect
}
\description{
\code{loopdataSubset} takes the loops and anchors present in dlo1
and uses the counts and samples from dlo2.
}
\details{
This function plays nice with \code{union} to ensure counts are correct
after taking the union of two loopdata objects. The subset function simply
returns the anchors and loops of dlo1 and the counts and colData of dlo2.
}
\examples{
# divide and recombine samples
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
naive <- jpn_chr1reg[,1:2]
primed <- jpn_chr1reg[,3:4]
np <- union(naive, primed)
# Subset from full to get correct counts
c.np <- loopdataSubset(np, jpn_chr1reg)
}

