% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{loopDistancePlot}
\alias{loopDistancePlot}
\alias{loopDistancePlot,loops-method}
\title{Visualize proportion of loops at distances}
\usage{
loopDistancePlot(dlo)

\S4method{loopDistancePlot}{loops}(dlo)
}
\arguments{
\item{dlo}{A loops object}
}
\value{
A ggplot2 plot
}
\description{
\code{loopDistancePlot} takes a loops object plots the individual samples
based on the proportion of PET counts at various distances
}
\details{
Distances for the loops are taken from the \code{rowData} slot.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
p1 <- loopDistancePlot(loops.small)
}

