% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffloop-class.R
\docType{class}
\name{loopfit-class}
\alias{loopfit}
\alias{loopfit-class}
\title{A class to represent ChIA-PET interaction data and an edgeR fit.}
\description{
A class to represent ChIA-PET interaction data and an edgeR fit.
}
\section{Slots}{

\describe{
\item{\code{loopdata}}{A loopdata with anchors, loops, counts, and colData}

\item{\code{fit}}{An edgeR fit from running the loopFit function}
}}

