% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{loopGenes}
\alias{loopGenes}
\alias{loopGenes,loops,GRanges-method}
\title{Determine genes contained within loops}
\usage{
loopGenes(dlo, genesGR)

\S4method{loopGenes}{loops,GRanges}(dlo, genesGR)
}
\arguments{
\item{dlo}{A loops object}

\item{genesGR}{A GRanges object of genes with mcol 'id'}
}
\value{
A matrix of comma separated gene names
}
\description{
\code{loopGenes} determines all gene bodies partially or fully contained
in a loop.
}
\details{
Function that annotates all loops. 'NA' if looping between chromosomes.
Otherwise, all gene names that are contained within a loop. 'None' if no
genes are in the loop body. If there are multiple, the function returns a
comma separated list. The length of the object returned by this function
should be the same length as the number of rows in the \code{loops} slot.
}
\examples{
# Determine the genes housed in the loops from our example
genes <- getHumanGenes()
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
loops.small <- loopGenes(loops.small,genes)

}

