% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffloop-class.R
\docType{class}
\name{loops-class}
\alias{loops}
\alias{loops-class}
\title{A class to represent ChIA-PET interaction data and annotations}
\description{
A class to represent ChIA-PET interaction data and annotations
}
\section{Slots}{

\describe{
\item{\code{anchors}}{A GRanges object describing loop anchor locations}

\item{\code{interactions}}{A matrix. Each row is an interaction between two anchors}

\item{\code{counts}}{A matrix with the number paired-end reads per loop per sample}

\item{\code{colData}}{A data.frame with features (colums) for each sample (rows)}

\item{\code{rowData}}{A data.frame with features (columns) for each loop (rows)}
}}

