% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loops.small}
\alias{loops.small}
\title{chr1:36000000-36300000 loops}
\format{A small loopdata object 
\describe{
  \item{anchors}{GRanges object of anchor locations}
  \item{loops}{indexes of interactions }
  \item{samples}{Two replicates each of jurkat, naive, and primed cells}
  \item{colData}{Groups identifying cell type and unnormalized sizeFactors}
  \item{rowData}{Base initialization with only loopWidth values}
  ...
}}
\source{
subsetRegion(loops,GRanges(c('1'),IRanges(c(36000000),c(36300000))))
}
\usage{
loops.small
}
\value{
A loops object
}
\description{
A loops object containing unique 108 loops with 27 anchors
for 6 samples and corresponding colData/rowData
}
\keyword{datasets}

