% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_io.R
\docType{methods}
\name{loopsMake.mango}
\alias{loopsMake.mango}
\alias{loopsMake.mango,ANY-method}
\title{Read preprocessed ChIA-PET data from mango}
\usage{
loopsMake.mango(beddir, samples = NA, mergegap = 500, ext = "fdr")

\S4method{loopsMake.mango}{ANY}(beddir, samples = NA, mergegap = 500,
  ext = "fdr")
}
\arguments{
\item{beddir}{A string. The preprocessed data directory; Required}

\item{samples}{A character vector. Optional list of samples to read in}

\item{mergegap}{An integer value of the radius to merge anchors; default 500}

\item{ext}{Specificies 'all' or 'fdr' file format; default 'fdr'}
}
\value{
A loops object where 'chr' is removed from the anchors.
}
\description{
\code{loopsMake.mango} reads in a data directory created by the
\code{mango} preprocessing pipeline and returns a loops object
}
\details{
This function reads in preprocessed ChIA-PET data produced by the
\code{mango} preprocessing pipeline. The \code{samples} argument specifies
which samples are read. If \code{samples} is not specified all samples will
be read. The \code{ext} specifies which type of file to look for, either
\code{all} or \code{fdr}, with \code{fdr} being the default. Under the
default, all samples with the extension \code{.fdr.mango} will be processed.
Finally, the \code{FDR} parameter (default = 1) specifies the minimum threshold
for loops to be added to the greater \code{loops} object. Currently, we do not
support importing the verbose output, so the verbose parameter when executing
mango should be set to FALSE or the user will have ot parse the file before
reading into diffloop using \code{awk}, \code{cut}, or something similar.
}
\examples{
# UPDATE THIS
bd<- system.file('extdata', 'esc_jurkat', package='diffloopdata')
}

