% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_io.R
\docType{methods}
\name{loopsMake}
\alias{loopsMake}
\alias{loopsMake,ANY-method}
\title{Read preprocessed ChIA-PET data from dnaloop}
\usage{
loopsMake(beddir, samples = NA, mergegap = 0, type = "all")

\S4method{loopsMake}{ANY}(beddir, samples = NA, mergegap = 0,
  type = "all")
}
\arguments{
\item{beddir}{A string. The preprocessed data directory}

\item{samples}{A character vector. Optional list of samples to read in}

\item{mergegap}{An integer value of the radius to merge anchors; default 0}

\item{type}{Specificies 'intra', 'inter', or 'all' looping. Default 'all'}
}
\value{
A loops object
}
\description{
\code{loopsMake} reads in a data directory created by the
\code{dnaloop} preprocessing pipeline and returns a loops object
}
\details{
This function reads in preprocessed ChIA-PET data produced by the
\code{dnaloop} preprocessing pipeline. The \code{samples} argument specifies
which samples are read. If \code{samples} is not specified all samples will
be read. The \code{type} option restricts loops whether they are on the same 
'inter' or different 'intra' chormosome. Default is 'all'.
}
\examples{
# Reading in all samples, no mergegap, all loops
bd<- system.file('extdata', 'esc_jurkat', package='diffloopdata')
# loops <- loopsMake(bd) #standard call

# Reading in a subset of samples, 1kb mergegap, only intrachromosomal
# looping
samples <- c('naive_esc_1', 'naive_esc_2')
naive.intra <- loopsMake(bd, samples, 1000, 'intra')

}

