% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\docType{methods}
\name{loopsSubset}
\alias{loopsSubset}
\alias{loopsSubset,loops,loops-method}
\title{Subset two difloop objects}
\usage{
loopsSubset(dlo1, dlo2)

\S4method{loopsSubset}{loops,loops}(dlo1, dlo2)
}
\arguments{
\item{dlo1}{A loops object}

\item{dlo2}{A loops object}
}
\value{
A loops obect
}
\description{
\code{loopsSubset} takes the interactions and anchors present in dlo1
and uses the counts and samples from dlo2.
}
\details{
This function plays nice with \code{union} to ensure counts are correct
after taking the union of two loops objects. The subset function simply
returns the anchors and interactions of dlo1 and the counts and colData of dlo2.
}
\examples{
# divide and recombine samples
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
naive <- loops.small[,1:2]
primed <- loops.small[,3:4]
np <- union(naive, primed)
# Subset from full to get correct counts
c.np <- loopsSubset(np, loops.small)
}

