% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffloop-class.R
\docType{class}
\name{looptest-class}
\alias{looptest}
\alias{looptest-class}
\title{A class to represent ChIA-PET interaction data and an association
test for the loops.}
\description{
A class to represent ChIA-PET interaction data and an association
test for the loops.
}
\section{Slots}{

\describe{
\item{\code{loopdata}}{A loopdata object with anchors, loops, counts, and colData}

\item{\code{results}}{A data.frame of association values per loop in the loopdata}
}}

