% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\docType{methods}
\name{loopTest}
\alias{loopTest}
\alias{loopTest,loopfit,missing,missing,missing-method}
\alias{loopTest,loopfit,missing,numeric,missing-method}
\alias{loopTest,loopfit,numeric,missing,missing-method}
\title{Compute looptest object}
\usage{
loopTest(y, coef = 2, contrast, method = "QLF")

\S4method{loopTest}{loopfit,missing,missing,missing}(y, coef = 2, contrast,
  method = "QLF")

\S4method{loopTest}{loopfit,numeric,missing,missing}(y, coef = 2, contrast,
  method = "QLF")

\S4method{loopTest}{loopfit,missing,numeric,missing}(y, coef = 2, contrast,
  method = "QLF")
}
\arguments{
\item{y}{A loopfit object for association}

\item{coef}{Specifies coefficient of design matrix}

\item{contrast}{Specifies comparison of groups from design matrix}

\item{method}{Specifies association method; only QLF is currently supported}
}
\value{
A looptest object
}
\description{
\code{loopTest} takes a \code{loopfit} object from the 
\code{loopFit} function and creates a \code{looptest} object.
}
\details{
This function returns a \code{looptest} object, which combines the results
from an association with the \code{loopdata}. The default association is
using coefficient 2 from the model matrix (e.g. good for pair
comparisons) but the user may specify a different coefficient. Currently, 
'QLF' is the only supported method, but new features may be added in
later developments. Users may also specify the contrast between the columns
in the design matrix as used in \code{edgeR}.
}
\examples{
# Differential loop fit
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
jpn_loopfit <- loopFit(jpn_chr1reg)
# Differential loop calling between naive and jurkat
assoc_jn <- loopTest(jpn_loopfit, coef = 2)
}

