% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{loopWidth}
\alias{loopWidth}
\alias{loopWidth,loops-method}
\title{Loop widths}
\usage{
loopWidth(dlo)

\S4method{loopWidth}{loops}(dlo)
}
\arguments{
\item{dlo}{A loops object}
}
\value{
An integer vector
}
\description{
\code{loopWidth} returns the width of a loop, which is defined as the
distance between the anchors containing a loop
}
\details{
This function returns a positive integer value of the number of basepairs
that separate two loops. If they are on separate chromosomes, it still
returns a value, but it will be non-sensical, so consider subsetting to
only intrachromosomal loops. Also, self-loops will return a postive number
that is the inter-anchor width. These loops should be handled using the 
removeSelfLoops() function.
}
\examples{
# Return the width for loops 
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
w <- loopWidth(loops.small)

}

