% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoFunctions.R
\docType{methods}
\name{mangoCorrection}
\alias{mangoCorrection}
\alias{mangoCorrection,ANY-method}
\title{Perform mango bias correction}
\usage{
mangoCorrection(lo, FDR = 1, PValue = 1, nbins = 10)

\S4method{mangoCorrection}{ANY}(lo, FDR = 1, PValue = 1, nbins = 10)
}
\arguments{
\item{lo}{A loops object.}

\item{FDR}{Minimum FDR value for loop to be included; default 1}

\item{PValue}{Minimum p0value for loop to be included; default 1}

\item{nbins}{Number of bins for mango computation}
}
\value{
A loops object where loops are filtered using mango bias correction
}
\description{
\code{mangoCorrection} takes a loops object and filters loops based
on the binomial model used in the mango ChIA-PET pipeline.
}
\details{
This function processes ChIA-PET data in a loops object and filters loops
that may be biased due to proximity or low PET counts as previously 
described by the \code{mango} pipeline. PET and anchor counts are aggregated
across all samples to compute statistical significance. Consider using a larger
number of bins (e.g. 30) for a larger data object when possible.
}
\examples{
rda <- paste(system.file("rda", package = "diffloop"), "loops.small.rda", sep = "/")
load(rda)
loops.small <- removeSelfLoops(loops.small)
loops.small.mango <- mangoCorrection(loops.small, PValue = 0.05)

}

