% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{manyLoopPlots}
\alias{manyLoopPlots}
\alias{manyLoopPlots,loops,GRanges-method}
\title{Plot several loop regions}
\usage{
manyLoopPlots(x, y, organism = "h", geneinfo = "NA", colorLoops = FALSE,
  cache = TRUE, maxCounts = -1, showAnchorWidths = FALSE)

\S4method{manyLoopPlots}{loops,GRanges}(x, y, organism = "h",
  geneinfo = "NA", colorLoops = FALSE, cache = TRUE, maxCounts = -1,
  showAnchorWidths = FALSE)
}
\arguments{
\item{x}{loops object}

\item{y}{GRanges object with many regions to be visualized}

\item{organism}{'h' for human or 'm' for mouse supported}

\item{geneinfo}{A data.frame manually specifying annotation (see Examples)}

\item{colorLoops}{Differentiates loops based on loop.type in loops object}

\item{cache}{logic variable (default = TRUE) to use gene annotation from July.2015 freeze}

\item{maxCounts}{Number of counts associated with thickest loop. Default is largest count
in region displayed}

\item{showAnchorWidths}{Display the width of the anchor on the plot? Default = FALSE}
}
\value{
Prints a time stamped .pdf file of top loops
}
\description{
\code{manyLoopPlots} takes a loops object and creates a time-stamped .pdf
file with loop plots (one per page) of all regions specified in the GRanges object.
}
\details{
Each plot will show one region sequentially that is supplied in the GRanges object.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
jpn.u <- removeSelfLoops(loops.small)
jpn_loopfit <- loopFit(jpn.u)
assoc_jn <- loopTest(jpn_loopfit, coef = 2)
#manyLoopPlots(assoc_jn, regs) #define regs as multiple GRanges

}

