% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{pcaPlot}
\alias{pcaPlot}
\alias{pcaPlot,loops-method}
\title{Visualize sample relationships}
\usage{
pcaPlot(dlo)

\S4method{pcaPlot}{loops}(dlo)
}
\arguments{
\item{dlo}{A loops object}
}
\value{
A ggplot2 plot
}
\description{
\code{pcaPlot} takes a loops object plots the individual samples
based on the principal components of the loop counts matrix
}
\details{
Groups for the principal component plots are derived from \code{colData}
and the normalizing factors are also taken from \code{colData}. While some
loops objects may have non-informative groups or size factors, they 
should always be present.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
p1 <- pcaPlot(loops.small)
}

