% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{plotTopLoops}
\alias{plotTopLoops}
\alias{plotTopLoops,looptest-method}
\title{Plot the most significant loops}
\usage{
plotTopLoops(lto, n = 0, PValue = 1, FDR = 1, organism = "h",
  colorLoops = FALSE)

\S4method{plotTopLoops}{looptest}(lto, n = 0, PValue = 1, FDR = 1,
  organism = "h", colorLoops = FALSE)
}
\arguments{
\item{lto}{looptest object}

\item{n}{number of loops to print (can remain 0 to specify from other parameters)
determined by PValue}

\item{PValue}{Maximum pvalue threshold for loop inclusion when printing loop plot}

\item{FDR}{False discovery rate threshold for inclusion}

\item{organism}{Either 'm' for mouse or 'h' for human.}

\item{colorLoops}{Default FALSE; specify true if results slot contains
loop.type from annotateLoops to visualize plots with varying colors for
CTCF looping and enhancer-promoter looping}
}
\value{
Prints a time stamped .pdf file of top loops
}
\description{
\code{plotTopLoops} takes a looptest object and creates a time-stamped .pdf
file with loop plots (one per page) of the top loops.
}
\details{
Each plot will show the region +/- 1 loopwidth of the loop with annotation specified
for either human or mouse. Assumes columns Pvalue and FDR are specified in the looptest
object. We recommend removing self loops before using this function (and in reality, 
before any association testing was called.)
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
jpn.u <- removeSelfLoops(jpn_chr1reg)
jpn_loopfit <- loopFit(jpn.u)
assoc_jn <- loopTest(jpn_loopfit, coef = 2)
plotTopLoops(assoc_jn, n=4)

}

