% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\docType{methods}
\name{quickAssocVoom}
\alias{quickAssocVoom}
\alias{quickAssocVoom,loops-method}
\title{Perform quick differential loop calling}
\usage{
quickAssocVoom(y)

\S4method{quickAssocVoom}{loops}(y)
}
\arguments{
\item{y}{A loops object for association}
}
\value{
A loops object
}
\description{
\code{quickAssocVoom} takes a loops object and performs a basic
\code{voom} association on the counts matrix and groups from \code{colData}
}
\details{
This function returns the output of fitting an \code{voom} model using
the groups defined in \code{colData} for the specific loops
object. The factor normalization is based on the \code{voom} model.
For quick association, the number of groups is restricted to two. If
a more complex group structure exists, consider using the \code{loopFit}
and \code{loopTest} functions.

If sizeFactors are present in the \code{loops} object, then those will
be used for normalization. If they are all 1s, new ones will be estimated
using the voom framework.
}
\examples{
# Differential loop calling between naive and primed
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
np <- loops.small[,1:4]
assoc_np_voom <- quickAssocVoom(np)
}

