% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{removeRegion}
\alias{removeRegion}
\alias{removeRegion,loops,GRanges-method}
\title{Remove region from loops object}
\usage{
removeRegion(dlo, region)

\S4method{removeRegion}{loops,GRanges}(dlo, region)
}
\arguments{
\item{dlo}{A loops object to be subsetted}

\item{region}{A GRanges object containing region of interest}
}
\value{
A loops object with no anchors touching the region given
}
\description{
\code{removeRegion} takes a \code{loops} object and a \code{GRanges}
object and returns a \code{loops} object where neither anchors map inside
the \code{GRanges} coordinates.
}
\examples{
# Remove region chr1:36000000-36100000
library(GenomicRanges)
regA <- GRanges(c('1'),IRanges(c(36000000),c(36100000)))
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
# Get rid of loop if either anchor touches that region
restricted <- removeRegion(loops.small, regA)
}

