% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{rmchr}
\alias{rmchr}
\alias{rmchr,GRanges-method}
\alias{rmchr,loops-method}
\title{Remove 'chr' from GRanges seqnames}
\usage{
rmchr(dlo)

\S4method{rmchr}{loops}(dlo)

\S4method{rmchr}{GRanges}(dlo)
}
\arguments{
\item{dlo}{A loops object or GRanges object}
}
\value{
An identical loops/GRanges object except 'chr' removed
}
\description{
\code{rmchr} takes a loops object or GRanges object and 
simply removes the 'chr' from seqnames, if is present
}
\details{
Often times, performing functions on GRanges objects can go awry 
if the seqnames are systematically different. A common example of
this is when some GRanges objects has the format of 'chr1' 
while the other has '1'. We can remove 'chr' from the first object
}
\examples{
library(GenomicRanges)
regA <- GRanges(c('1'),IRanges(c(36200000),c(36300000)))
addchr(regA)
regA
rmchr(regA)
regA
}

