% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\docType{methods}
\name{sampleNames,loopdata-method}
\alias{sampleNames,loopdata-method}
\alias{sampleNames<-,loopdata,ANY-method}
\title{Grab/Update Sample Names}
\usage{
\S4method{sampleNames}{loopdata}(object)

\S4method{sampleNames}{loopdata,ANY}(object) <- value
}
\arguments{
\item{object}{A loopdata object}

\item{value}{New names when using set replace}
}
\value{
Vector of sample names
}
\description{
\code{sampleNames} takes a loopdata object returns the names of 
the samples in the structure. One can also update the names using
set replace.
}
\details{
The examples show both accession and updating sample names.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
sampleNames(jpn_chr1reg)
nnames <- c("one", "two", "three", "four", "five", "six")
sampleNames(jpn_chr1reg) <- nnames
}

