% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\docType{methods}
\name{sampleNames,loops-method}
\alias{sampleNames,loops-method}
\alias{sampleNames<-,loops,ANY-method}
\title{Grab/Update Sample Names}
\usage{
\S4method{sampleNames}{loops}(object)

\S4method{sampleNames}{loops,ANY}(object) <- value
}
\arguments{
\item{object}{A loops object}

\item{value}{New names when using set replace}
}
\value{
Vector of sample names
}
\description{
\code{sampleNames} takes a loops object returns the names of 
the samples in the structure. One can also update the names using
set replace.
}
\details{
The examples show both accession and updating sample names.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
sampleNames(loops.small)
nnames <- c('one', 'two', 'three', 'four', 'five', 'six')
sampleNames(loops.small) <- nnames
}

