% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\docType{methods}
\name{slidingWindowTest}
\alias{slidingWindowTest}
\alias{slidingWindowTest,looptest,numeric,numeric-method}
\title{Combined association test for all loops in a defined region}
\usage{
slidingWindowTest(x, window, step)

\S4method{slidingWindowTest}{looptest,numeric,numeric}(x, window, step)
}
\arguments{
\item{x}{A looptest object}

\item{window}{The length a window will be for combined association}

\item{step}{The size that the window will shift for each association}
}
\value{
A data.frame sorted by FDR
}
\description{
\code{slidingWindowTest} takes a \code{looptest} object and 
integer values of the association window and the distance between 
consecutive windows.
}
\details{
This function returns a data.frame sorted by FDR of each region. The engine
loops over each chromosome and defines the first window at the left-most
loop and slides the window right until no more loops are present in \code{x}
Each region is determined from a sliding window of fixed length.
The combined significance measure per feature is computed via the Simes 
method for intrachromosomal loops where at least one anchor from the loop
overlaps with the region.
}
\examples{
# Sliding window test 100kb at a time between naive and jurkat
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
jpn_loopfit <- loopFit(jpn_chr1reg)
# Differential loop calling between naive and jurkat
assoc_jn <- loopTest(jpn_loopfit, coef = 2)
sw_jn <- slidingWindowTest(assoc_jn, 100000, 100000)

}

