% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\docType{methods}
\name{splitSamples}
\alias{splitSamples}
\alias{splitSamples,loops-method}
\title{Split samples into their own loops object}
\usage{
splitSamples(dlo)

\S4method{splitSamples}{loops}(dlo)
}
\arguments{
\item{dlo}{A loops object}
}
\value{
A list of loops objects with one sample per index.
}
\description{
\code{splitSamples} takes a loops object and returns a list of loops
objects where each sample populates its own loops object
}
\details{
This function splits the colData and counts slots for each sample
but makes copies of the anchors, interactions, and rowdata
}
\examples{
# Updating groups from all 'group1' to meaningful designations
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
split <- splitSamples(loops.small)

}

