% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{[,loopdata,numeric,numeric,missing-method}
\alias{[,loopdata,missing,numeric,missing-method}
\alias{[,loopdata,numeric,missing,missing-method}
\alias{[,loopdata,numeric,numeric,missing-method}
\title{Extract parts of a loopdata object}
\usage{
\S4method{[}{loopdata,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{loopdata,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{loopdata,numeric,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A loopdata object for subsetting}

\item{i}{Loops to be subsetted}

\item{j}{Samples to be subsetted}

\item{drop}{Other non-essential parameters needed for sub}
}
\value{
A loopdata object
}
\description{
Extract parts of a loopdata object
}

