% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{[,loops,numeric,numeric,missing-method}
\alias{[,loops,missing,numeric,missing-method}
\alias{[,loops,numeric,missing,missing-method}
\alias{[,loops,numeric,numeric,missing-method}
\title{Extract parts of a loops object}
\usage{
\S4method{[}{loops,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{loops,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{loops,numeric,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A loops object for subsetting}

\item{i}{Loops to be subsetted}

\item{j}{Samples to be subsetted}

\item{drop}{Other non-essential parameters needed for sub}
}
\value{
A loops object
}
\description{
Extract parts of a loops object
}

