% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{[,looptest,numeric,missing,missing-method}
\alias{[,looptest,numeric,missing,missing-method}
\title{Extract parts of looptest}
\usage{
\S4method{[}{looptest,numeric,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A looptest object for subsetting}

\item{i}{Loops to be subsetted}

\item{j}{NULL}

\item{drop}{NULL}
}
\value{
A looptest object
}
\description{
Can only subsample loops, so only the i parameter can be handled
}

