% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{subsetLoops}
\alias{subsetLoops}
\alias{subsetLoops,loopdata,logical-method}
\alias{subsetLoops,loopdata,numeric-method}
\title{Subset loops}
\usage{
subsetLoops(dlo, idxa)

\S4method{subsetLoops}{loopdata,logical}(dlo, idxa)

\S4method{subsetLoops}{loopdata,numeric}(dlo, idxa)
}
\arguments{
\item{dlo}{A loopdata object}

\item{idxa}{A numeric vector or logical vector}
}
\value{
A loopdata object
}
\description{
\code{subsetLoops} restricts the loops and counts matrix to only those 
specified by \code{idxa}, either numerically or logically
}
\details{
This function returns a loopdata object where the loops are retained only
if they meet a logical criteria or are included in the numeric vector of
\code{idxa}. Only the anchors that reference a loop in the subsetted 
loopdata object are retained.
}
\examples{
# Return the first 10 loops
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
#' ten <- subsetLoops(jpn_chr1reg, 1:10)

# Subset loops with widths greater than 10000
big <- subsetLoops(jpn_chr1reg, loopWidth(jpn_chr1reg) >= 10000)

}

