% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{subsetRegion}
\alias{subsetRegion}
\alias{subsetRegion,loopdata,GRanges-method}
\title{Extract region from loopdata object}
\usage{
subsetRegion(dlo, region)

\S4method{subsetRegion}{loopdata,GRanges}(dlo, region)
}
\arguments{
\item{dlo}{A loopdata object to be summarized}

\item{region}{A GRanges object containing region of interest}
}
\value{
A loopdata object
}
\description{
\code{subsetRegion} takes a \code{loopdata} object and a \code{GRanges}
object and returns a \code{loopdata} object where both anchors map inside
the \code{GRanges} coordinates
}
\details{
This function returns a data.frame where the left and right anchors are 
visualized together along with the loop width, individual counts, and
any anchor meta-data that has been annotated into the anchors \code{GRanges}
object.
}
\examples{
# Grab region chr1:36000000-36100000
library(GenomicRanges)
regA <- GRanges(c('1'),IRanges(c(36000000),c(36100000)))
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
jpn_chr1reg.small <- subsetRegion(jpn_chr1reg, regA)
}

