% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruan.R
\docType{methods}
\name{subsetRegionAB}
\alias{subsetRegionAB}
\alias{subsetRegionAB,loops,GRanges,GRanges-method}
\title{Retain loops that have anchors in two specified regions}
\usage{
subsetRegionAB(dlo, regionA, regionB)

\S4method{subsetRegionAB}{loops,GRanges,GRanges}(dlo, regionA, regionB)
}
\arguments{
\item{dlo}{A loops object}

\item{regionA}{A GRanges object}

\item{regionB}{A GRanges object}
}
\value{
A loops object
}
\description{
\code{subsetRegionAB} returns a loops object where one anchor 
maps to regionA and the other maps to region B
}
\examples{
# Return the width for loops 
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
regA <- GRanges(c('1'),IRanges(c(36000000),c(36100000)))
regB <- GRanges(c('1'),IRanges(c(36200000),c(36300000)))
splits <- subsetRegionAB(dlo, regA, regB)

}

