% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{summarize,loopdata-method}
\alias{summarize,loopdata-method}
\alias{summarize,looptest-method}
\title{Link the anchors and loops back together}
\usage{
\S4method{summarize}{loopdata}(.data)

\S4method{summarize}{looptest}(.data)
}
\arguments{
\item{.data}{A loopdata object to be summarized}
}
\value{
A data.frame
}
\description{
\code{summarize} takes a \code{loopdata/looptest} object and breaks the 
loopdata structure resulting in a \code{data.frame}.
}
\details{
This function returns a \code{data.frame} where the left and right anchors 
are visualized together along with the loop width, individual counts, and
any anchor meta-data that has been annotated into the anchors GRanges
object. When a \code{looptest} object is the input, the additional rows 
in the \code{results} slot is added on but the individual counts are 
excluded.
}
\examples{
# Summarizing the first ten loops in \\code{jpn_chr1reg}
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
summarydf <- summarize(jpn_chr1reg[1:10,])
# Summarizing the loops and significance results between naive and primed
summarylt <- summarize(quickAssoc(jpn_chr1reg[,1:4])[1:10,])
}

