% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{summary,loops-method}
\alias{summary,loops-method}
\title{Link the anchors and interactions back together}
\usage{
\S4method{summary}{loops}(object)
}
\arguments{
\item{object}{A loops object to be summarized}
}
\value{
A data.frame
}
\description{
\code{summary} takes a \code{loops} object and breaks the 
loop data structure resulting in a \code{data.frame}.
}
\details{
This function returns a \code{data.frame} where the left and right anchors 
are visualized together along with the loop width, individual counts, and
any anchor meta-data that has been annotated into the anchors GRanges
object as well as any rowData varianble
}
\examples{
# Summarizing the first ten loops in \\code{loops.small}
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
summarydf <- summary(loops.small[1:10,])
# Summarizing the loops and significance results between naive and primed
summarylt <- summary(quickAssoc(loops.small[,1:4])[1:10,])
}

