% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{tail,loopdata-method}
\alias{tail,loopdata-method}
\title{Extract last part of loopdata}
\usage{
\S4method{tail}{loopdata}(x, n = 6, ...)
}
\arguments{
\item{x}{A loopdata object}

\item{n}{Number of lines to view}

\item{...}{Other non-essential params}
}
\value{
A loopdata object
}
\description{
Extract last part of loopdata
}

