% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\docType{methods}
\name{tail,looptest-method}
\alias{tail,looptest-method}
\title{Extract last part of looptest}
\usage{
\S4method{tail}{looptest}(x, n = 6, ...)
}
\arguments{
\item{x}{A looptest object}

\item{n}{Number of lines to view}

\item{...}{Other non-essential params}
}
\value{
A looptest object
}
\description{
Extract last part of looptest
}

