% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\docType{methods}
\name{topLoops}
\alias{topLoops}
\alias{topLoops,loops,missing,numeric-method}
\alias{topLoops,loops,numeric,missing-method}
\alias{topLoops,loops,numeric,numeric-method}
\title{Grab top loops}
\usage{
topLoops(dlo, FDR, PValue)

\S4method{topLoops}{loops,numeric,numeric}(dlo, FDR, PValue)

\S4method{topLoops}{loops,numeric,missing}(dlo, FDR, PValue)

\S4method{topLoops}{loops,missing,numeric}(dlo, FDR, PValue)
}
\arguments{
\item{dlo}{A loops object}

\item{FDR}{Maximum threshold for False Discovery Rate; default = 1}

\item{PValue}{Maximum threshold for P-value; default = 1}
}
\value{
A loops object subsetted by specified parameters
}
\description{
\code{topLoops} takes a loops object and performs basic filtering
for \code{FDR} or \code{PValue}
}
\details{
This function returns a subsetted \code{loops} object where all
loops meet the significance threshold specificed by the parameters
in the function call
}
\examples{
# Differential loop calling between naive and primed
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
np <- loops.small[,1:4]
assoc_np <- quickAssoc(np)
top_np <- topLoops(assoc_np, FDR = 0.3)
}

