% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\docType{methods}
\name{union,loopdata,loopdata-method}
\alias{union,loopdata,loopdata-method}
\title{Combine two loopdata objects}
\usage{
\S4method{union}{loopdata,loopdata}(x, y)
}
\arguments{
\item{x}{A loopdata object}

\item{y}{A loopdata object}
}
\value{
A loopdata obect
}
\description{
\code{union} combines two loopdata objects' loops and anchors and 
populates the \code{colData} matrix where available
}
\details{
This function returns a single loopdata object that has all the 
anchors and loops contained in the two loopdata objects that were
part of the input. However, when the two objects have different samples,
the counts matrix will contain missing values (e.g. when loop counts
in x are not in y, those values are unknown). While the number of loops,
colData, and anchors should be correct, we need to correct the counts
using a subsetting function.
}
\examples{
# divide and recombine samples
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
naive <- jpn_chr1reg[,1:2]
primed <- jpn_chr1reg[,3:4]
np <- union(naive, primed)
# Subset from full to get correct counts
c.np <- loopdataSubset(np, jpn_chr1reg)
}

