% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\docType{methods}
\name{union,loops,loops-method}
\alias{union,loops,loops-method}
\title{Combine two loops objects}
\usage{
\S4method{union}{loops,loops}(x, y)
}
\arguments{
\item{x}{A loops object}

\item{y}{A loops object}
}
\value{
A loops obect
}
\description{
\code{union} combines two loops objects' interactions and anchors and 
populates the \code{colData} matrix where available
}
\details{
This function returns a single loops object that has all the 
anchors and interactions contained in the two loops objects that were
part of the input. However, when the two objects have different samples,
the counts matrix will contain missing values (e.g. when loop counts
in x are not in y, those values are unknown). While the number of interactions,
colData, and anchors should be correct, we need to correct the counts
using a subsetting function. The row data gets re-initialized here to only
the loop widths
}
\examples{
# divide and recombine samples
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
naive <- loops.small[,1:2]
primed <- loops.small[,3:4]
np <- union(naive, primed)
# Subset from full to get correct counts
c.np <- loopsSubset(np, loops.small)
}

