% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{intrachromosomal}
\alias{intrachromosomal}
\alias{intrachromosomal,loopdata-method}
\title{Loops within chromosomes}
\usage{
intrachromosomal(dlo)

\S4method{intrachromosomal}{loopdata}(dlo)
}
\arguments{
\item{dlo}{A loopdata object}
}
\value{
A loopdata object where all loops are on the same chromosome.
}
\description{
\code{intrachromosomal} restricts loops to those where anchors are observed
on the same chromosomes
}
\details{
This function subsets the \code{loopdata} object into only those loops that
have both anchors on the same chromosome
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)

# Compute number of loops on same chromosome
dim(intrachromosomal(jpn_chr1reg))
samechromo <- intrachromosomal(jpn_chr1reg)

}

