% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_io.R
\docType{methods}
\name{loopdataMake}
\alias{loopdataMake}
\alias{loopdataMake,ANY-method}
\title{Read preprocessed ChiA-PET data}
\usage{
loopdataMake(beddir, samples = NA, mergegap = 0, type = "all")

\S4method{loopdataMake}{ANY}(beddir, samples = NA, mergegap = 0,
  type = "all")
}
\arguments{
\item{beddir}{A string. The preprocessed data directory}

\item{samples}{A character vector. Optional list of samples to read in}

\item{mergegap}{An integer value of the radius to merge anchors; default 0}

\item{type}{Specificies "intra", "inter", or "all" looping. Default "all"}
}
\value{
A loopdata object
}
\description{
\code{loopdataMake} reads in a data directory created by the
\code{dnaloop} preprocessing pipeline and returns a loopdata object
}
\details{
This function reads in preprocessed ChIA-PET data produced by the
\code{dnaloop} preprocessing pipeline. The preprocessed directory 
contains one subdirectory per sample. The \code{samples} argument specifies
which samples are read. if \code{samples} is not specified all samples will
be read. \code{type} restricts loops whether they are on the same "inter" or
different "intra" chormosome. Default is "all"
}
\examples{
# Reading in all samples, no mergegap, all loops
bd<- system.file('extdata', 'esc_jurkat', package='diffloopdata')
loops <- loopdataMake(bd)

# Reading in a subset of samples, 1kb mergegap, only intrachromosomal
# looping
samples <- c('naive_esc_1', 'naive_esc_2')
naive.intra <- loopdataMake(bd, samples, 1000, "intra")

}

