% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{loopGenes}
\alias{loopGenes}
\alias{loopGenes,loops,GRanges-method}
\alias{loopGenes,loops,missing-method}
\title{Determine genes contained within loops}
\usage{
loopGenes(dlo, genesGR)

\S4method{loopGenes}{loops,GRanges}(dlo, genesGR)

\S4method{loopGenes}{loops,missing}(dlo, genesGR)
}
\arguments{
\item{dlo}{A loops object}

\item{genesGR}{A GRanges object of genes in first mcol.}
}
\value{
A matrix of comma separated gene names
}
\description{
\code{loopGenes} determines all gene bodies partially or fully contained
in a loop.
}
\details{
Function that annotates all loops. If there are multiple, the function returns a
comma separated list. Adds a "loopGenes" column to the rowData slot. If
the genesGR is left blank, diffloop will use protein coding genes for human
from hg19.
}
\examples{
# Determine the genes housed in the loops from our example
genes <- getHumanGenes()
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
loops.small <- loopGenes(loops.small,genes)

}

