% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{loopMetrics}
\alias{loopMetrics}
\alias{loopMetrics,loopdata-method}
\title{Types of loops}
\usage{
loopMetrics(dlo)

\S4method{loopMetrics}{loopdata}(dlo)
}
\arguments{
\item{dlo}{A loopdata object}
}
\value{
A data.frame
}
\description{
\code{loopMetrics} counts number of loops for each sample and returns
whether they are single, self, unique, or none
}
\details{
This function shows the number of loops for each sample based on four
types. Single refers to having only one anchor for a the loop whereas
none has no unique anchors. If using the \code{loopdataMake} pipeline, only
self and unique loops will be observed when running this function
}
\examples{
# Return loop metrics for number of each type for each sample 
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
loopMetrics(jpn_chr1reg)

}

