% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{loopWidth}
\alias{loopWidth}
\alias{loopWidth,loopdata-method}
\title{Loop widths}
\usage{
loopWidth(dlo)

\S4method{loopWidth}{loopdata}(dlo)
}
\arguments{
\item{dlo}{A loopdata object}
}
\value{
An integer vector
}
\description{
\code{loopWidth} returns the width of a loop, which is defined as the
distance between the anchors containing a loop
}
\details{
This function returns a positive integer value of the number of basepairs
that separate two loops. If they are on separate chromosomes, it still
returns a value, but it will be non-sensical, so consider subsetting to
only intrachromosomal loops.
}
\examples{
# Return the width for loops 
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
w <- loopWidth(jpn_chr1reg)

}

