% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{numAnchors}
\alias{numAnchors}
\alias{numAnchors,loops-method}
\title{Get number of anchors in each sample}
\usage{
numAnchors(x)

\S4method{numAnchors}{loops}(x)
}
\arguments{
\item{x}{A loops object to be summarized}
}
\value{
A data.frame of each sample and the number of anchors
}
\description{
\code{numAnchors} takes a \code{loops} object and a summarizes the
number of anchors that support all the interactions (count >= 1) in the object
}
\details{
This function returns a data.frame where the column names specify the
sample in the original \code{loops} object and the only row shows
the number of anchors used to support that sample
}
\examples{
# Show number of anchors each sample is supported by
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
numAnchors(loops.small)
}

