% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{numLoops}
\alias{numLoops}
\alias{numLoops,loops,missing-method}
\alias{numLoops,loops,numeric-method}
\title{Per-sample loop quantities}
\usage{
numLoops(dlo, nloops = 1:10)

\S4method{numLoops}{loops,numeric}(dlo, nloops = 1:10)

\S4method{numLoops}{loops,missing}(dlo, nloops = 1:10)
}
\arguments{
\item{dlo}{A loops object}

\item{nloops}{A numeric vector of counts to be considered}
}
\value{
A data.frame
}
\description{
\code{numLoops} counts number of loops for each sample based on the
index of nloops and returns a data.frame
}
\details{
This function shows the number of unique loops with at least
\code{nloops} in counts. Can be used to quickly visualize relative
sequencing depth between samples
}
\examples{
# Determine what samples have loops with 1-20 counts
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
nLoops <- numLoops(loops.small, 1:20)

# Determine what samples loops with 1-10 counts by default
nLoops <- numLoops(loops.small)
}

