% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{methods}
\name{pcaPlot}
\alias{pcaPlot}
\alias{pcaPlot,loopdata-method}
\alias{pcaPlot,looptest-method}
\title{Visualize sample relationships}
\usage{
pcaPlot(dlo)

\S4method{pcaPlot}{loopdata}(dlo)

\S4method{pcaPlot}{looptest}(dlo)
}
\arguments{
\item{dlo}{A loopdata/looptest object}
}
\value{
A ggplot2 plot
}
\description{
\code{pcaPlot} takes a loopdata object plots the individual samples
based on the principal components of the loop counts matrix
}
\details{
Groups for the principal component plots are derived from \code{colData}
and the normalizing factors are also taken from \code{colData}. While some
loopdata objects may have non-informative groups or size factors, they 
should always be present.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
p1 <- pcaPlot(jpn_chr1reg)
}

