% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopFunctions.R
\docType{methods}
\name{removeSelfLoops}
\alias{removeSelfLoops}
\alias{removeSelfLoops,loopdata-method}
\title{Remove self loops}
\usage{
removeSelfLoops(dlo)

\S4method{removeSelfLoops}{loopdata}(dlo)
}
\arguments{
\item{dlo}{A loopdata object}
}
\value{
A loopdata object
}
\description{
\code{removeSelfLoops} removes instances where a loop is observed
between the same anchor
}
\details{
This function removes loops from the \code{loops} slot that reference
the same index of the \code{anchors} slot.
}
\examples{
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
jpn_unique <- removeSelfLoops(jpn_chr1reg)

}

