% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{methods}
\name{subsetRegion}
\alias{subsetRegion}
\alias{subsetRegion,loops,GRanges,missing-method}
\alias{subsetRegion,loops,GRanges,numeric-method}
\title{Extract region from loops object}
\usage{
subsetRegion(dlo, region, nanchors)

\S4method{subsetRegion}{loops,GRanges,numeric}(dlo, region, nanchors)

\S4method{subsetRegion}{loops,GRanges,missing}(dlo, region, nanchors)
}
\arguments{
\item{dlo}{A loops object to be subsetted}

\item{region}{A GRanges object containing region of interest}

\item{nanchors}{Number of anchors to be contained in GRanges object. Default 2}
}
\value{
A loops object
}
\description{
\code{subsetRegion} takes a \code{loops} object and a \code{GRanges}
object and returns a \code{loops} object where both anchors map inside
the \code{GRanges} coordinates by default. Once can specify where only
one anchor is in the region instead.
}
\details{
By default, \code{nachors = 2}, meaning both anchors need to be in the 
region for the loop to be preserved when extracting. However, by specifying
a numeric 1, interactions with either the left or right anchor will be extracted.
Loops with both anchors in the region will be excluded (exclusive \code{or}).
To get an inclusive \code{or}, take the union of subsetting both with 1 and 2.
}
\examples{
# Grab region chr1:36000000-36100000
library(GenomicRanges)
regA <- GRanges(c('1'),IRanges(c(36000000),c(36100000)))
rda<-paste(system.file('rda',package='diffloop'),'loops.small.rda',sep='/')
load(rda)
# Both anchors in region
loops.small.two <- subsetRegion(loops.small, regA)
#Only one anchor in region
loops.small.one <- subsetRegion(loops.small, regA, 1)
#Either one or two anchors in region
loops.small.both <- union(loops.small.one, loops.small.two)
}

