% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\docType{methods}
\name{updateLDGroups}
\alias{updateLDGroups}
\alias{updateLDGroups,loopdata-method}
\title{Update groups in colData for loopdata object}
\usage{
updateLDGroups(dlo, groups)

\S4method{updateLDGroups}{loopdata}(dlo, groups)
}
\arguments{
\item{dlo}{A loopdata object}

\item{groups}{A character vector. Lists the groups each sample belongs in}
}
\value{
A loopdata object with new groups in \code{colData}
}
\description{
\code{updateLDGroups} changes the \code{groups} column in \code{colData} for
a \code{loopdata} object
}
\details{
This function updates the \code{groups} column in \code{colData} for
a \code{loopdata} object. Make sure that the length of \code{groups}
the number of samples in \code{colData}!
}
\examples{
# Updating groups from all 'group1' to meaningful designations
rda<-paste(system.file('rda',package='diffloop'),'jpn_chr1reg.rda',sep='/')
load(rda)
celltypes <- c('naive1','naive1','primed2','primed2','jurkat3','jurkat3')
jpn_chr1reg <- updateLDGroups(jpn_chr1reg, celltypes)

}

